requirements.txt
streamlit
requests
pillow